requirements.txt
streamlit>=1.32.0
mistralai>=1.6.0
loguru>=0.7.3
python-dotenv>=1.0.0
pyperclip>=1.8.2
pytest>=8.0.0
